+++ test/unit-tests/function/arithmetic/mod.test.js
@@ -34,8 +34,14 @@ describe('mod', function () {
     approx.equal(mod(-5, 3), 1)
   })
 
-  it('should throw an error if the divisor is negative', function () {
-    assert.throws(function () { mod(10, -4) })
+  it('should handle precise approximation of float approximation', function () {
+    approx.equal(mod(0.1, 0.01), 0)
+    approx.equal(mod(0.15, 0.05), 0)
+    approx.equal(mod(1.23456789, 0.00000000001), 0)
+  })
+
+  it('should calculate mod for negative divisor', function () {
+    assert.strictEqual(mod(10, -4), -2)
   })
 
   it('should throw an error if used with wrong number of arguments', function () {